fix Exception \Error has already been caught
ref:
   http://php.net/manual/en/class.error.php
   http://php.net/manual/en/class.exception.php
   http://php.net/manual/en/class.throwable.php
   https://github.com/walkor/Workerman/blob/master/Lib/Constants.php#L41

Since PHP 7, Error&Exception both implements throwable and are in the
same level. When class Error do not exist, Error class is defined and
extends Exception, so the Error should be caught in the first order
then Exception.

Used when catch clauses are not in the correct order (from the more
specific to the more generic) or one exception class is caught twice.
If you don't fix the order, some exceptions may not be catched by
the most specific handler